requirements.txt
ollama
langchain==0.3.7
langchain-community==0.3.5
langchain-ollama==0.2.1
pydantic==2.9.2
chromadb==0.5.23
rank-bm25==0.2.2
sentence-transformers==3.3.1
fastapi==0.115.6
uvicorn
APScheduler==3.10.4
python-dotenv
pyyaml
streamlit==1.40.2
plyer
win10toast
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
tenacity==9.0.0
sqlcipher3==0.5.3
keyring==25.5.0
cryptography==44.0.0
alembic==1.14.0
loguru==0.7.3
packaging==24.2
pytest
mypy
pip-audit
pyinstaller
google-auth
google-api-core